Refactor atlas page components
