+++ frontend/src/app/features/scripture-atlas/components/map/scripture-atlas-map.component.html
@@ -1,0 +1,27 @@
+<div class="map-area">
+  <div class="maps-container" [class.split]="splitView">
+    <div class="map-wrapper" [class.half]="splitView">
+      <div id="modern-map" class="map-instance"></div>
+      <div class="map-label" *ngIf="splitView">Modern View</div>
+    </div>
+
+    <div class="map-wrapper" [class.half]="splitView" *ngIf="splitView">
+      <div id="ancient-map" class="map-instance"></div>
+      <div class="map-label">Ancient View</div>
+    </div>
+  </div>
+
+  <app-scripture-atlas-nav
+    [currentCityIndex]="currentCityIndex"
+    [totalCities]="cities.length"
+    (previous)="previousCity.emit()"
+    (next)="nextCity.emit()">
+  </app-scripture-atlas-nav>
+
+  <app-scripture-atlas-playback-progress
+    *ngIf="isPlaying"
+    [currentIndex]="currentCityIndex"
+    [total]="cities.length"
+    [progress]="playbackProgress">
+  </app-scripture-atlas-playback-progress>
+</div>
