+++ frontend/src/app/features/scripture-atlas/components/map/scripture-atlas-map.component.scss
@@ -1,0 +1,126 @@
+@import '../../shared/variables';
+
+.map-area {
+  flex: 1;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  padding: 1rem;
+  background: $gray-50;
+}
+
+.maps-container {
+  flex: 1;
+  position: relative;
+  display: flex;
+  background: white;
+  border-radius: 0.75rem;
+  overflow: hidden;
+  box-shadow: 0 2px 4px 0 rgb(0 0 0 / 0.06);
+
+  &.split {
+    .map-wrapper {
+      &.half {
+        width: 50%;
+
+        &:first-child {
+          border-right: 2px solid $gray-300;
+        }
+      }
+    }
+  }
+}
+
+.map-wrapper {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.map-instance {
+  width: 100%;
+  height: 100%;
+}
+
+.map-label {
+  position: absolute;
+  top: 1rem;
+  left: 1rem;
+  padding: 0.375rem 0.75rem;
+  background: rgba(white, 0.9);
+  backdrop-filter: blur(4px);
+  border-radius: 0.375rem;
+  font-size: 0.75rem;
+  font-weight: 600;
+  color: $gray-700;
+  box-shadow: 0 2px 4px 0 rgb(0 0 0 / 0.1);
+}
+
+// Custom marker styles for the map
+:host ::ng-deep {
+  .leaflet-container {
+    background: $gray-100;
+    font-family: inherit;
+  }
+
+  .leaflet-control-zoom {
+    border: none;
+
+    a {
+      background: white !important;
+      border: 1px solid $gray-300 !important;
+      color: $gray-700 !important;
+      width: 32px !important;
+      height: 32px !important;
+      line-height: 30px !important;
+
+      &:hover {
+        background: $gray-50 !important;
+        color: $gray-900 !important;
+      }
+    }
+  }
+
+  .leaflet-control-attribution {
+    background: rgba(white, 0.8) !important;
+    font-size: 0.625rem;
+  }
+
+  .city-marker-wrapper {
+    z-index: 600 !important;
+  }
+
+  .custom-city-marker {
+    border: 3px solid;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-weight: 700;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2);
+    transition: all 0.2s;
+    position: relative;
+
+    &.selected {
+      animation: pulse 2s infinite;
+    }
+
+    .marker-content {
+      font-size: 1rem;
+    }
+  }
+
+  .route-arrow {
+    font-size: 20px;
+    font-weight: bold;
+    opacity: 0.7;
+  }
+}
+
+// Responsive
+@media (max-width: 768px) {
+  .map-area {
+    padding: 0.5rem;
+  }
+}
