+++ frontend/src/app/features/scripture-atlas/components/nav/scripture-atlas-nav.component.html
@@ -1,0 +1,14 @@
+<div class="nav-buttons">
+  <button class="nav-btn" (click)="previous.emit()" [disabled]="currentCityIndex === 0">
+    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor">
+      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 19l-7-7 7-7" />
+    </svg>
+    Previous
+  </button>
+  <button class="nav-btn primary" (click)="next.emit()">
+    Next
+    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor">
+      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 5l7 7-7 7" />
+    </svg>
+  </button>
+</div>
