+++ frontend/src/app/features/scripture-atlas/components/playback-progress/scripture-atlas-playback-progress.component.html
@@ -1,0 +1,17 @@
+<div class="playback-progress">
+  <svg viewBox="0 0 36 36">
+    <path
+      d="M18 2.0845 a 15.9155 15.9155 0 0 1 0 31.831 a 15.9155 15.9155 0 0 1 0 -31.831"
+      fill="none"
+      stroke="#e5e7eb"
+      stroke-width="3" />
+    <path
+      d="M18 2.0845 a 15.9155 15.9155 0 0 1 0 31.831 a 15.9155 15.9155 0 0 1 0 -31.831"
+      fill="none"
+      stroke="#6366f1"
+      stroke-width="3"
+      [attr.stroke-dasharray]="'100 100'"
+      [attr.stroke-dashoffset]="100 - progress" />
+  </svg>
+  <span class="progress-text">{{ currentIndex + 1 }}/{{ total }}</span>
+</div>
