+++ frontend/src/app/features/scripture-atlas/components/playback-progress/scripture-atlas-playback-progress.component.scss
@@ -1,0 +1,29 @@
+@import '../../shared/variables';
+
+.playback-progress {
+  position: absolute;
+  bottom: 1.5rem;
+  left: 1.5rem;
+  width: 60px;
+  height: 60px;
+  background: white;
+  border-radius: 50%;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 50;
+
+  svg {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    transform: rotate(-90deg);
+  }
+
+  .progress-text {
+    font-size: 0.75rem;
+    font-weight: 700;
+    color: $primary;
+  }
+}
