--- frontend/src/app/features/scripture-atlas/components/scripture-atlas-header/scripture-atlas-header.component.scss
@@ -1,44 +1,0 @@
-.atlas-header {
-  background: #fff;
-  border-bottom: 1px solid #e5e7eb;
-}
-
-.header-content {
-  display: flex;
-  justify-content: space-between;
-  padding: 8px;
-}
-
-.journey-select {
-  margin-left: 8px;
-}
-
-.header-controls {
-  display: flex;
-  gap: 8px;
-}
-
-.control-btn {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  border: none;
-  background: transparent;
-  cursor: pointer;
-}
-
-.control-btn.active {
-  color: #6366f1;
-}
-
-.info-badge {
-  display: inline-flex;
-  align-items: center;
-  gap: 4px;
-  margin-right: 8px;
-  font-size: 12px;
-}
-
-.info-badge.highlight {
-  color: #ef4444;
-}
