--- frontend/src/app/features/scripture-atlas/components/scripture-atlas-map/scripture-atlas-map.component.scss
@@ -1,32 +1,0 @@
-.map-area {
-  position: relative;
-  height: 100%;
-  width: 100%;
-}
-
-.maps-container {
-  position: relative;
-  height: calc(100% - 60px);
-  width: 100%;
-  display: flex;
-}
-
-.map-wrapper.half {
-  width: 50%;
-}
-
-.map-instance {
-  height: 100%;
-  width: 100%;
-}
-
-.map-label {
-  position: absolute;
-  top: 8px;
-  left: 8px;
-  background: rgba(0, 0, 0, 0.5);
-  color: #fff;
-  padding: 2px 6px;
-  font-size: 12px;
-  border-radius: 4px;
-}
