--- frontend/src/app/features/scripture-atlas/components/scripture-atlas-playback-progress/scripture-atlas-playback-progress.component.scss
@@ -1,25 +1,0 @@
-.playback-progress {
-  position: absolute;
-  bottom: 16px;
-  right: 16px;
-  width: 48px;
-  height: 48px;
-}
-
-.playback-progress svg {
-  width: 100%;
-  height: 100%;
-}
-
-.progress-text {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 12px;
-  font-weight: bold;
-}
