--- frontend/src/app/features/scripture-atlas/components/scripture-atlas-playback-progress/scripture-atlas-playback-progress.component.ts
@@ -1,33 +1,0 @@
-import { Component, Input } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'app-scripture-atlas-playback-progress',
-  standalone: true,
-  imports: [CommonModule],
-  template: `
-    <div class="playback-progress">
-      <svg viewBox="0 0 36 36">
-        <path
-          d="M18 2.0845 a 15.9155 15.9155 0 0 1 0 31.831 a 15.9155 15.9155 0 0 1 0 -31.831"
-          fill="none"
-          stroke="#e5e7eb"
-          stroke-width="3" />
-        <path
-          d="M18 2.0845 a 15.9155 15.9155 0 0 1 0 31.831 a 15.9155 15.9155 0 0 1 0 -31.831"
-          fill="none"
-          stroke="#6366f1"
-          stroke-width="3"
-          [attr.stroke-dasharray]="'100 100'"
-          [attr.stroke-dashoffset]="100 - progress" />
-      </svg>
-      <span class="progress-text">{{ currentIndex + 1 }}/{{ total }}</span>
-    </div>
-  `,
-  styleUrls: ['./scripture-atlas-playback-progress.component.scss']
-})
-export class ScriptureAtlasPlaybackProgressComponent {
-  @Input() currentIndex = 0;
-  @Input() total = 0;
-  @Input() progress = 0;
-}
