--- frontend/src/app/features/scripture-atlas/components/scripture-atlas-timeline/scripture-atlas-timeline.component.scss
@@ -1,63 +1,0 @@
-.timeline-bar {
-  width: 100%;
-  padding: 8px;
-}
-
-.timeline-track {
-  position: relative;
-  height: 40px;
-}
-
-.timeline-progress {
-  position: absolute;
-  top: 18px;
-  left: 0;
-  height: 4px;
-  background: #6366f1;
-}
-
-.timeline-markers {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 40px;
-}
-
-.timeline-marker {
-  position: absolute;
-  transform: translateX(-50%);
-  cursor: pointer;
-  text-align: center;
-}
-
-.marker-dot {
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-  background: #d1d5db;
-  margin: 0 auto;
-}
-
-.timeline-marker.visited .marker-dot {
-  background: #6366f1;
-}
-
-.timeline-marker.memorized .marker-dot {
-  background: #10b981;
-}
-
-.timeline-marker.active .marker-dot {
-  background: #ef4444;
-}
-
-.marker-label {
-  font-size: 10px;
-}
-
-.timeline-slider {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-}
