+++ frontend/src/app/features/scripture-atlas/components/sidebar/scripture-atlas-sidebar.component.scss
@@ -1,0 +1,278 @@
+@import '../../shared/variables';
+
+.sidebar {
+  width: $sidebar-width;
+  height: 100%;
+  background: transparent;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  transition: margin-left 0.3s ease;
+  padding: 1rem 0 1rem 1rem;
+
+  &.collapsed {
+    margin-left: -$sidebar-width;
+  }
+}
+
+.sidebar-wrapper {
+  background: white;
+  border: 1px solid $gray-200;
+  border-radius: 0.75rem;
+  box-shadow: 0 2px 4px 0 rgb(0 0 0 / 0.06);
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.sidebar-tabs {
+  display: flex;
+  background: $gray-50;
+  border-bottom: 1px solid $gray-200;
+  position: relative;
+}
+
+.tab-btn {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.375rem;
+  padding: 0.75rem;
+  background: transparent;
+  border: none;
+  border-bottom: 2px solid transparent;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: $gray-600;
+  cursor: pointer;
+  transition: all 0.2s;
+
+  svg {
+    width: 16px;
+    height: 16px;
+  }
+
+  &:hover {
+    color: $gray-800;
+    background: rgba($gray-100, 0.5);
+  }
+
+  &.active {
+    color: $primary;
+    border-bottom-color: $primary;
+    background: white;
+  }
+}
+
+.close-btn {
+  position: absolute;
+  top: 0.5rem;
+  right: 0.5rem;
+  width: 28px;
+  height: 28px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: $gray-100;
+  border: none;
+  border-radius: 0.375rem;
+  cursor: pointer;
+  transition: all 0.2s;
+
+  svg {
+    width: 16px;
+    height: 16px;
+    color: $gray-500;
+  }
+
+  &:hover {
+    background: $gray-200;
+  }
+}
+
+.sidebar-content {
+  flex: 1;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  @include custom-scrollbar;
+  overflow-y: auto;
+}
+
+.content-header {
+  margin-bottom: 1rem;
+
+  h2 {
+    font-size: 1.25rem;
+    font-weight: 700;
+    color: $gray-900;
+    margin: 0 0 0.25rem;
+  }
+
+  .verse-reference {
+    font-size: 0.875rem;
+    color: $primary;
+    margin: 0;
+  }
+
+  .city-modern {
+    font-size: 0.875rem;
+    color: $gray-600;
+    margin: 0;
+  }
+}
+
+.verse-text {
+  flex: 1;
+  font-size: 0.875rem;
+  line-height: 1.6;
+  color: $gray-700;
+  overflow-y: auto;
+
+  .verse-number {
+    font-weight: 700;
+    color: $primary;
+    margin-right: 0.25rem;
+  }
+
+  .verse {
+    margin-bottom: 0.5rem;
+  }
+}
+
+.city-description {
+  font-size: 0.875rem;
+  line-height: 1.5;
+  color: $gray-700;
+  margin-bottom: 1rem;
+}
+
+.city-events {
+  margin-bottom: 1rem;
+
+  h4 {
+    font-size: 0.75rem;
+    font-weight: 700;
+    text-transform: uppercase;
+    letter-spacing: 0.05em;
+    color: $gray-500;
+    margin: 0 0 0.5rem;
+  }
+
+  ul {
+    margin: 0;
+    padding-left: 1.25rem;
+
+    li {
+      font-size: 0.8125rem;
+      line-height: 1.5;
+      color: $gray-600;
+      margin-bottom: 0.25rem;
+    }
+  }
+}
+
+.city-fact {
+  display: flex;
+  gap: 0.75rem;
+  padding: 0.75rem;
+  background: rgba($warning, 0.05);
+  border: 1px solid rgba($warning, 0.2);
+  border-radius: 0.5rem;
+  margin-top: auto;
+
+  svg {
+    width: 20px;
+    height: 20px;
+    color: $warning;
+    flex-shrink: 0;
+  }
+
+  p {
+    font-size: 0.75rem;
+    line-height: 1.4;
+    color: $gray-700;
+    margin: 0;
+  }
+}
+
+.content-footer {
+  margin-top: auto;
+  padding-top: 1rem;
+  border-top: 1px solid $gray-200;
+  display: flex;
+  gap: 0.5rem;
+}
+
+.action-btn {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.375rem;
+  padding: 0.625rem;
+  background: white;
+  border: 1px solid $gray-300;
+  border-radius: 0.5rem;
+  font-size: 0.75rem;
+  font-weight: 500;
+  color: $gray-700;
+  cursor: pointer;
+  transition: all 0.2s;
+
+  svg {
+    width: 16px;
+    height: 16px;
+  }
+
+  &:hover {
+    background: $gray-50;
+    border-color: $gray-400;
+  }
+
+  &.primary {
+    background: $primary;
+    border-color: $primary;
+    color: white;
+
+    &:hover {
+      background: $primary-dark;
+      border-color: $primary-dark;
+    }
+
+    &.completed {
+      background: $success;
+      border-color: $success;
+    }
+  }
+
+  &.completed {
+    opacity: 0.8;
+  }
+}
+
+// Responsive
+@media (max-width: 1024px) {
+  .sidebar {
+    position: absolute;
+    z-index: 60;
+    box-shadow: 0 0 20px rgba(0, 0, 0, 0.1);
+    background: white;
+    padding: 0;
+    border-right: 1px solid $gray-200;
+
+    &.collapsed {
+      transform: translateX(-100%);
+      margin-left: 0;
+    }
+  }
+}
+
+@media (max-width: 768px) {
+  .sidebar {
+    width: 100%;
+    max-width: 320px;
+  }
+}
